requirements.txt
torch
torchvision
timm
pandas
numpy
Pillow
tqdm
scikit-learn
matplotlib
openpyxl
